Use vcr to record changelog lookups for better test coverage (#7)
Before, we used a `FakeChangelogLocator` when testing. This meant that
the logic for looking up changelogs, which is messy and buggy, wasn't
being tested.

This PR installs `vcr` and `webmock` gems, and updates the existing
`Report` test to use the real changelog lookup algorithm involving HTTP
requests.

The resulting vcr cassettes are checked into Git, so that test runs are
fast.