Fix a bug preventing CHANGELOG.md files from being discovered (#8)
When `update-interactive` searches for a changelog of a particular gem
release, it first checks the `changelog_uri` metadata in the gem itself
via the rubygems.org API.

Failing that, if the gem is hosted on GitHub, it scans the GitHub
repository for a file name matching
`changelog|changes|history|news|release`.

However, due to a bug, this match was case-sensitive, causing it to miss
the common `CHANGELOG` spelling. Oops!

Fix this mistake and add a vcr-based test coverage.