Remove thor dependency; use optparse directly (#14)
Thor is useful for handling complex sets of options, arguments, and
sub-commands. For this project it is overkill. This PR removes thor so
that we have one less dependency to manage.

In place of thor, I created a `CLI::Options` class that encapsulates the
option parsing logic, with corresponding unit tests.