Change all/reverse key bindings to unmodified a/r (#18)
Currently, `ctrl-a` and `ctrl-r` are used for select-all and
reverse-selection. We get these as default key bindings from
[tty-prompt](https://github.com/piotrmurach/tty-prompt), the library we
use for the multi-select menu. As discussed in #16, these defaults are a
bit problematic.

This PR changes the key bindings to `a` and `r` instead (i.e. without
the control-key modifier).

I also added test coverage for these, as well as the `j`/`k` and `n`/`p`
bindings added in #4.

Fixes #16