Refactor `Reporter` → `Updater`
Before, the logic for finding outdated gems and updating those gems was
split across 2 classes: `Reporter` and `Report`. This refactor
consolidates those concerns into a single `Updater` class. As a result,
the `Report` class becomes simpler, and more of a DTO.

The motivation is to create a single extension point (i.e. `Updater`)
for the update logic to be customized for the upcoming `--latest`
feature.