Add openssl gem to dev dependencies to fix cert errors (#96)
Integration tests that fetched changelogs from certain HTTPS servers
recently started failing with errors like this:

> OpenSSL::SSL::SSLError: SSL_connect returned=1 errno=0
peeraddr=151.101.193.227:443 state=error: certificate verify failed
(unable to get certificate CRL)

This is a known issue with OpenSSL + Ruby and the fix is to upgrade the
openssl gem. I've added the gem to the development dependencies.