Gracefully handle OpenSSL errors when fetching changelogs (#97)
The Ruby OpenSSL implementation has a bug that prevents it from
connecting to certain HTTPS servers. This can cause `bundle ui` to crash
with `OpenSSL::SSL::SSLError` while trying to fetch changelogs.

As a workaround, this commit gracefully handles the error so that the
changelog lookup is skipped, rather than crashing altogether.

Ultimately the solution is outside the scope this project; users will
need to update the `openssl` gem to the latest version. If a user hasn't
yet done so, this commit ensures `bundle ui` will at least work without
crashing.